Complete redo
I know I said last update that I was keeping the model as it was, but truth is that I wasn't really understanding it, and now I have completely redone it, and now it should function as it needs to, as well as just being entirely better. This does mean that I have completely restarted my java stuff too, although there was hardly anything there to start with. Hopefully from here I can start coding and branching and completing a lot more.